fix: fix due to drive letter inconsistencies
