+++ packages/vscode-wdio-test/src/repository.ts
@@ -319,7 +319,7 @@ export class TestRepository extends MetadataRepository implements TestRepository
         for (const [key, value] of this._fileMap.entries()) {
             // The path of the Spec file is the third one, as it is the next level after Workspace,WdioConfig.
             const candidatePath = key.split(TEST_ID_SEPARATOR)[2]
-            if (normalizedSpecFilePath === candidatePath) {
+            if (normalizedSpecFilePath === normalizePath(candidatePath)) {
                 return value
             }
         }
